Broke out CRUDder interface
Had to add to autoload to make it load interfaces
